refactoring di opposite
